rlezhaiko/lesson_24/requirements.txt
asgiref==3.6.0
Django==4.1.5
django-ckeditor==6.5.1
django-js-asset==2.0.0
django-seed==0.3.1
Faker==16.7.0
psycopg2-binary==2.9.5
python-dateutil==2.8.2
six==1.16.0
sqlparse==0.4.3
toposort==1.9
